component wise tensor product

+++ test_vector.cpp
@@ -12,10 +12,10 @@ TEST_CASE("Vector dot product", "[vector]") {
 }
 
 TEST_CASE("Vector cross product (3D)", "[vector]") {
-    vector<int, 3> v1(1, 2, 3);
-    vector<int, 3> v2(4, 5, 6);
-    vector<int, 3> product = cross(v1, v2);
-    vector<int, 3> expected_product(-3, 6, -3);
+    vector<float, 3> v1(1.0f, 2.0f, 3.0f);
+    vector<float, 3> v2(4.0f, 5.0f, 6.0f);
+    vector<float, 3> product = cross(v1, v2);
+    vector<float, 3> expected_product(-3.0f, 6.0f, -3.0f);
     REQUIRE(product == expected_product);
 }
